requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-rest-swagger
drf_yasg
django-cors-headers